Added 'preciseDecimal' parser.
Default `decimal` parser rounds to tenth of a degree Celsius. For users converting to Fahrenheit, the full 1000'ths precision can be used and the rounding done after conversion. For example 19.250 C would be rounded to 19.3 C.  If converted to F there is a 0.1 degree difference between these values, which may be significant depending on the application.